Change from absolute to relative paths
This commit changes the paths to be relative. Thus, the input files can
be executed on any computer independent of the folder structure outside
the repository. Note that the simulation must be run from within the
`input_files` folder.

Additionally, the loading of the spent fuel composition has been
adapted: the `allow_pickle` argument has been set to `True`, else
the compositions do not load.